regional conditionality for standard spreads
now each region can have TRUE/FALSE for each of the "stardard" spreads we have from global book.

Currently this conditionality is implementend at the spread level for clarity. Therefore if one wants to change a plot on a spread then a new spread will be defined (and created in the partX)

Next to implement conditionality for cover page AND for subsetting the data (countries) etc. within chunkcs